Moved routes/db setup to app.js so index only starts listen. For testing reasons
